refactor: extract renderJobCreate helper function - Extract `renderJobCreate` function into a separate file - Update test files to use the new `renderJobCreate` helper function
